+++ spec/acceptance/gem_update_spec.rb
@@ -20,16 +20,16 @@ RSpec.describe GemUpdater do
       * i18n 1.14.1 → 1.14.7
       [changelog](https://github.com/ruby-i18n/i18n/releases)
 
-      * json 2.6.3 → 2.10.2
-      [changelog](https://github.com/ruby/json/blob/master/CHANGES.md#2025-03-12-2102)
+      * json 2.6.3 → 2.12.0
+      [changelog](https://github.com/ruby/json/blob/master/CHANGES.md#2025-05-12-2120)
 
       * minitest 5.20.0 → 5.25.5
       [changelog](https://github.com/minitest/minitest/blob/master/History.rdoc#5255--2025-03-12-)
 
-      * parallel 1.23.0 → 1.26.3
+      * parallel 1.23.0 → 1.27.0
 
-      * parser 3.2.2.3 → 3.3.7.4
-      [changelog](https://github.com/whitequark/parser/blob/v3.3.7.4/CHANGELOG.md)
+      * parser 3.2.2.3 → 3.3.8.0
+      [changelog](https://github.com/whitequark/parser/blob/v3.3.8.0/CHANGELOG.md)
 
       * racc 1.7.1 → 1.8.1
       [changelog](https://github.com/ruby/racc/releases)
@@ -37,11 +37,11 @@ RSpec.describe GemUpdater do
       * regexp_parser 2.6.1 → 2.10.0
       [changelog](https://github.com/ammar/regexp_parser/blob/master/CHANGELOG.md#2100---2024-12-25---janosch-müller)
 
-      * rubocop 1.38.0 → 1.75.1
-      [changelog](https://github.com/rubocop/rubocop/releases/tag/v1.75.1)
+      * rubocop 1.38.0 → 1.75.6
+      [changelog](https://github.com/rubocop/rubocop/releases/tag/v1.75.6)
 
-      * rubocop-ast 1.29.0 → 1.43.0
-      [changelog](https://github.com/rubocop/rubocop-ast/blob/master/CHANGELOG.md#1430-2025-03-25)
+      * rubocop-ast 1.29.0 → 1.44.1
+      [changelog](https://github.com/rubocop/rubocop-ast/blob/master/CHANGELOG.md#1441-2025-04-11)
 
       * unicode-display_width 2.4.2 → 3.1.4
       [changelog](https://github.com/janlelis/unicode-display_width/blob/main/CHANGELOG.md#314)